+++ app/src/main/java/com/example/samplecomposeapp/list/lazyColumn/LazyColumn.kt
@@ -12,7 +12,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.samplecomposeapp.list.CustomItem
+import com.example.samplecomposeapp.list.components.CustomItem
 import com.example.samplecomposeapp.list.model.Person
 import com.example.samplecomposeapp.list.repository.PersonRepository
